Add files via upload
This project is about Python and R scripts that import, manipulate, and export employee-related data. The Python code was used to manipulate data and execute the main functions of extracting employee data and exporting the extracted data into a zip file. On the other hand, the R script was used to unzip and print the data.
Below are the requirements used to write and execute the program:
(A)	 Python: In writing the codes in python environment, python 3.x was used to install the required libraries (pandas, csv and the zipfile).
To install dependencies using Python, execute the following commands:
1.	Clone the repository
2.	Open the Jupyter Notebook file (‘assignment.ipy’) in Jupyter Notebook.
3.	Import data by loading the zipped total.csv.zip file containing employee data to a Pandas DataFrame.
4.	Import panda as pd.
5.	Run the cells to execute the code.
6.	Create the ‘Employee Profile.zip’ folder that will contain the employee details in csv (Comma-Separated Values).
7.	Create the employee function. Define the function to fetch employee details based on the name of an employee.
8.	Create a dictionary to process employee details
9.	Error Handling: Carefully handle any probable errors in case of data processing.
10.	Export the details of an employee to a CSV and zip it
11.	Open the R Studio file.
12.	Unzip the file by extracting the contents of the Total.csv.
13.	Load the library.
14.	Finally, display the data using the print command.
Python handled processing the data and exporting while R handled the unzipping and viewing processes.